[DT-405] Move apps related code to the apps module
