+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/presentation/CategoryAppsViewModel.kt
@@ -1,8 +1,8 @@
-package cm.aptoide.pt.feature_categories.presentation
+package cm.aptoide.pt.feature_apps.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import cm.aptoide.pt.feature_categories.domain.CategoryAppsUseCase
+import cm.aptoide.pt.feature_apps.domain.CategoryAppsUseCase
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.catch
