+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/presentation/ViewModelProvider.kt
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.feature_appview.presentation
+package cm.aptoide.pt.feature_apps.presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.runtime.Composable
@@ -7,8 +7,9 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
-import cm.aptoide.pt.feature_appview.domain.AppInfoUseCase
-import cm.aptoide.pt.feature_appview.domain.AppVersionsUseCase
+import cm.aptoide.pt.feature_apps.domain.CategoryAppsUseCase
+import cm.aptoide.pt.feature_apps.domain.AppInfoUseCase
+import cm.aptoide.pt.feature_apps.domain.AppVersionsUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
@@ -16,6 +17,7 @@ import javax.inject.Inject
 class InjectionsProvider @Inject constructor(
   val appInfoUseCase: AppInfoUseCase,
   val appVersionsUseCase: AppVersionsUseCase,
+  val categoryAppsUseCase: CategoryAppsUseCase,
 ) : ViewModel()
 
 @Composable
@@ -49,6 +51,23 @@ fun appVersionsViewModel(packageName: String): AppVersionsViewModel {
         return AppVersionsViewModel(
           appVersionsUseCase = injectionsProvider.appVersionsUseCase,
           packageName = packageName,
+        ) as T
+      }
+    }
+  )
+}
+
+@Composable
+fun categoryAppsViewModel(categoryName: String): CategoryAppsViewModel {
+  val injectionsProvider = hiltViewModel<InjectionsProvider>()
+  return viewModel(
+    key = categoryName,
+    factory = object : ViewModelProvider.Factory {
+      override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        @Suppress("UNCHECKED_CAST")
+        return CategoryAppsViewModel(
+          categoryName = categoryName,
+          categoryAppsUseCase = injectionsProvider.categoryAppsUseCase,
         ) as T
       }
     }
