+++ feature_categories/src/main/java/cm/aptoide/pt/feature_categories/presentation/ViewModelProvider.kt
@@ -6,14 +6,12 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
 import cm.aptoide.pt.feature_categories.data.CategoriesRepository
-import cm.aptoide.pt.feature_categories.domain.CategoryAppsUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class InjectionsProvider @Inject constructor(
   val categoriesRepository: CategoriesRepository,
-  val categoryAppsUseCase: CategoryAppsUseCase,
 ) : ViewModel()
 
 @Composable
@@ -27,23 +25,6 @@ fun CategoriesViewModel(requestUrl: String): CategoriesViewModel {
         return CategoriesViewModel(
           categoriesWidgetUrl = requestUrl,
           categoriesRepository = injectionsProvider.categoriesRepository,
-        ) as T
-      }
-    }
-  )
-}
-
-@Composable
-fun categoryAppsViewModel(categoryName: String): CategoryAppsViewModel {
-  val injectionsProvider = hiltViewModel<InjectionsProvider>()
-  return viewModel(
-    key = categoryName,
-    factory = object : ViewModelProvider.Factory {
-      override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        @Suppress("UNCHECKED_CAST")
-        return CategoryAppsViewModel(
-          categoryName = categoryName,
-          categoryAppsUseCase = injectionsProvider.categoryAppsUseCase,
         ) as T
       }
     }
